pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tokenwise-llm"
version = "0.5.0"
description = "Intelligent LLM task planner — decompose tasks, route to optimal models, enforce budgets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "TokenWise Contributors" }]
keywords = ["llm", "router", "planner", "ai", "openai", "budget"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/itsarbit/tokenwise"
Repository = "https://github.com/itsarbit/tokenwise"
Issues = "https://github.com/itsarbit/tokenwise/issues"
Changelog = "https://github.com/itsarbit/tokenwise/blob/master/CHANGELOG.md"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "mypy",
    "ruff",
    "respx",
    "types-pyyaml>=6.0.12.20250915",
]
benchmark = ["matplotlib"]

[project.scripts]
tokenwise = "tokenwise.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/tokenwise"]

[tool.hatch.build]
artifacts = ["*.json"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
